refactor: js 폴더에 파일 이동
