+++ src/js/controller/LottoSimulator.js
@@ -1,10 +1,10 @@
 import { LOTTO_CONSTANT, LOTTO_RANKING, PRINT_MESSAGE } from '../data/constants.js';
-import InputView from '../view/InputView.js';
+import InputView from '../view/InputView'
 import OutputView from '../view/OutputView.js';
 import LottoUtils from '../domain/LottoUtils.js';
 import Lotto from '../domain/Lotto.js';
 import WinningLotto from '../domain/WinningLotto.js';
-import Console from '../utils/Console.js';
+import Console from '../utils/Console';
 
 class LottoSimulator {
   #lottos;
