+++ src/js/domain/Lotto.js
