+++ src/js/domain/LottoUtils.js
