+++ src/js/utils/Validator.js
