+++ src/js/view/InputView.js
@@ -3,8 +3,7 @@ import Console from '../utils/Console.js';
 const InputView = {
   async readUserInput(message) {
     return await Console.read(message);
-  }
-
+  },
 };
 
 export default InputView;
